% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{COLORS}
\alias{COLORS}
\title{Motor insurance data}
\format{
vectors with 7 colors
}
\usage{
data(frenchmotor)
}
\description{
Dataset from CASDataset R package
}
\details{
variables are
\itemize{
\item 0 #5BBCD6 light blue (from Wes Anderson color palette)
\item 1 #FF0000 red
\item A #00A08A green (from Wes Anderson color palette)
\item B #F2AD00 yellow (from Wes Anderson color palette)
\item with #046C9A dark blue
\item without #C93312 dark red
\item 2 #0B775E dark green
}
}
\examples{
data(COLORS)
}
\keyword{colors}
