% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density_score}
\alias{density_score}
\title{Beta-kernel density estimation of a score}
\usage{
density_score(y)
}
\arguments{
\item{y}{The vector of scores}
}
\value{
A function that is the density estimate
}
\description{
Estimate the density of the score defined on $[0,1]$
}
\examples{
d1 <- density_score(runif(100));
d2 <- density_score(rbeta(100,2,1));
}
