% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_transport}
\alias{draw_transport}
\title{Optimal transport plot for scores}
\usage{
draw_transport(
  yA,
  yB,
  densA,
  densB,
  limA = c(0, 1),
  limB = c(0, 1),
  limY = c(0, 12),
  lab = c("A", "B"),
  sub = 5
)
}
\arguments{
\item{yA}{vector of predictions in group A}

\item{yB}{vector of predictions in group B}

\item{densA}{kernel density estimate associated with sample yA}

\item{densB}{kernel density estimate associated with sample yB}

\item{limA}{range of values for yA}

\item{limB}{range of values for yB}

\item{limY}{magnitude of the density}

\item{lab}{names for colors (default A and B)}

\item{sub}{number of subdivisions of scores axis}
}
\value{
A plot
}
\description{
Plots the optimal transport between two wamples
}
\examples{
NA
}
