% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fair_metrics}
\alias{fair_metrics}
\title{Fairness metrics}
\usage{
fair_metrics(
  threshold = 0.5,
  name_outcome = "y",
  outcome_ref = "1",
  base = toydata2,
  name_sensitive = "sensitive",
  sensitive_ref = "B",
  pred_outcome = probs
)
}
\arguments{
\item{threshold}{a model}

\item{name_outcome}{xxxxx}

\item{outcome_ref}{xxxxx}

\item{base}{xxxxx}

\item{name_sensitive}{xxxxx}

\item{sensitive_ref}{xxxxx}

\item{pred_outcome}{xxxxx}
}
\value{
a vector with 11 metrics, and details (table and graphs) on those 1.  metrics
}
\description{
Compute different fairness metrics
}
\examples{
data(toydata2)
model_gam = mgcv::gam(y~s(x1,k=12,bs="cr")+s(x2,k=12,bs="cr")+s(x3,k=12,bs="cr"),
data=toydata2,family=binomial)
pred_y = predict(model_gam, type="response")
fair_metrics(pred_outcome = pred_y)$metrics
}
