% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{frenchmotor}
\alias{frenchmotor}
\title{Motor insurance data}
\format{
data frame with 12437 obs. of  18 variables
}
\usage{
data(frenchmotor)
}
\description{
Dataset from CASDataset R package
}
\details{

}
\examples{
data(frenchmotor)
}
\keyword{dataset}
