% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{germancredit}
\alias{germancredit}
\title{German credit data}
\format{
data frame
}
\usage{
data(germancredit)
}
\description{
Dataset from CASDataset R package
}
\details{
See also \href{http://cas.uqam.ca/}{CASdatasets}. Variables are
\itemize{
\item Account_status          : Factor w/ 4 levels "< 0 DM",">= 200 DM",..: 1 3 4 1 1 4 4 3 4 3 ...
\item Duration                : num  6 48 12 42 24 36 24 36 12 30 ...
\item Credit_history          : Factor w/ 5 levels "all credits at this bank paid back duly",..: 2 4 2 4 3 4 4 4 4 2 ...
\item Purpose                 : Factor w/ 10 levels "business","car (new)",..: 8 8 5 6 2 5 6 3 8 2 ...
\item Credit_amount           : num  1169 5951 2096 7882 4870 ...
\item Savings_bonds           : Factor w/ 5 levels "< 100 DM",">= 1000 DM",..: 5 1 1 1 1 5 4 1 2 1 ...
\item Present_employment_since: Factor w/ 5 levels "< 1 year",">= 7 years",..: 2 3 4 4 3 3 2 3 4 5 ...
\item Installment_rate        : num  4 2 2 2 3 2 3 2 2 4 ...
\item Other_debtors_guarantors: Factor w/ 3 levels "co-applicant",..: 3 3 3 2 3 3 3 3 3 3 ...
\item Resident_since          : num  4 2 3 4 4 4 4 2 4 2 ...
\item Property                : Factor w/ 4 levels "building society savings agreement / life insurance",..: 3 3 3 1 4 4 1 2 3 2 ...
\item Age                     : num  67 22 49 45 53 35 53 35 61 28 ...
\item Other_installment_plans : Factor w/ 3 levels "bank","none",..: 2 2 2 2 2 2 2 2 2 2 ...
\item Housing                 : Factor w/ 3 levels "rent","own","for free": 2 2 2 3 3 3 2 1 2 2 ...
\item Existing_credits        : num  2 1 1 1 2 1 1 1 1 2 ...
\item Job                     : Factor w/ 4 levels "management / self-employed / highly qualified employee / officer",..: 2 2 4 2 2 4 2 1 4 1 ...
\item People_maintenance_for  : num  1 1 2 2 2 2 1 1 1 1 ...
\item Telephone               : Factor w/ 2 levels "none","yes": 2 1 1 1 1 2 1 2 1 1 ...
\item Foreign_worker          : Factor w/ 2 levels "no","yes": 2 2 2 2 2 2 2 2 2 2 ...
\item y                       : Factor w/ 2 levels "BAD","GOOD": 2 1 2 2 1 2 2 2 2 1 ...
\item sensitive               : Factor w/ 2 levels "Female","Male": 1 2 1 1 1 1 1 1 1 1 ...
}
}
\examples{
data(germancredit)
}
\keyword{dataset}
