% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OPT}
\alias{OPT}
\title{Compute the transport function}
\usage{
OPT(y0, y1, u = 0)
}
\arguments{
\item{y0}{The vector of initial vector}

\item{y1}{The vector of final vector}

\item{u}{Value at which transport is applied}
}
\value{
The transported value
}
\description{
Function for (univariate) optimal transport, see also \href{https://en.wikipedia.org/wiki/Transportation_theory_(mathematics)#Optimal_transportation_on_the_real_line}{wikipedia}
}
\examples{
str(OPT(y0 = rnorm(100,0,1),
        y1 = rnorm(100,1,2), 
        u = 0)
}
