% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_transport}
\alias{plot_transport}
\title{Compute the transport function on a grid}
\usage{
plot_transport(y0, y1, u = seq(min(y0), max(y0), length = 251))
}
\arguments{
\item{y0}{The vector of initial vector}

\item{y1}{The vector of final vector}

\item{u}{The grid to compute values for the transport plot}
}
\value{
A dataframe with $x$ and $y$ used in the transport function plot
}
\description{
Compute values use to plot the (univariate) optimal transport
}
\examples{
str(plot_transport(y0 = rnorm(100,0,1),
                   y1 = rnorm(100,1,2), 
                   u = seq(-3,3,by=.1))
}
\seealso{
\code{\link[=OPT]{OPT()}}
}
