% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{telematic}
\alias{telematic}
\title{Telematic data}
\format{
data frame with 1177 obs. of  16 variables
}
\usage{
data(telematic)
}
\description{
Telematic dataset (real data)
}
\details{
variables are
\itemize{
\item  id                : chr  "1" "2" "3" "4" ...
\item  Total_Distance    : num  3649 9066 27156 8558 17853 ...
\item  Drive_Score       : num  67 63 74 77 68 84 71 16 61 80 ...
\item  Time_Day          : num  109 116 73 108 106 63 108 136 117 105 ...
\item  Style_Score       : num  85 92 55 74 65 32 91 16 64 33 ...
\item  Corner_Score      : num  86 93 61 67 61 32 92 19 30 27 ...
\item  Acceleration_Score: num  88 80 49 91 89 40 94 16 51 42 ...
\item  Braking_Score     : num  84 93 64 82 72 46 90 21 87 50 ...
\item  Total_Night_Time  : num  34 28 24 25 38 20 23 27 38 21 ...
\item  Total_Time        : num  79 339 464 437 649 437 349 24 523 500 ...
\item  Acceleration      : Factor w/ 2 levels "High","Low": 2 1 2 1 1 2 1 2 2 2 ...
\item  Brake             : Factor w/ 2 levels "High","Low": 2 1 1 1 1 2 1 2 1 1 ...
\item  Corner            : Factor w/ 2 levels "High","Low": 1 1 1 1 1 2 1 2 2 2 ...
\item  insured_gender    : Factor w/ 2 levels "F","M": 2 1 1 2 2 2 2 2 2 1 ...
\item  age               : num  21 20 30 52 21 34 22 29 22 53 ...
\item  claim             : Factor w/ 2 levels "no","yes": 1 1 1 1 1 1 1 1 2 2 ...
}
}
\examples{
data(telematic)
table(telematic$insured_gender, telematic$claim)
}
\keyword{dataset}
