% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy1.R
\docType{data}
\name{toy1}
\alias{toy1}
\title{Toy1 Dataset}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
toy1
}
\description{
Simple toy dataset with 20 observations
}
\examples{
data(toy1)
str(toy1)

}
\keyword{datasets}
