% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{toydata1}
\alias{toydata1}
\title{Toy data no1}
\format{
data frame with 1000 obs. of  3 variables
}
\usage{
data(toydata1)
}
\description{
Toy dataset
}
\details{
variables are
\itemize{
\item x: (continuous) legitimate explanatory variable: num  -1.05838 -0.75016 1.54972 0.00442 -1.14538 ...
\item s: (binary) sensitive attribute: Factor w/ 2 levels "B","A": 1 2 2 2 2 2 1 1 2 1 ...
\item y: (binary) outome: Factor w/ 2 levels "0","1": 1 1 2 1 2 2 2 1 1 1 ...
}
}
\examples{
data(toydata1)
data(toydata1_train)
data(toydata1_valid)
}
\keyword{dataset}
