% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{toydata2}
\alias{toydata2}
\title{Toy data no2}
\format{
data frame with 1000 obs. of  6 variables
}
\usage{
data(toydata2)
}
\description{
Toy dataset
}
\details{
variables are
\itemize{
\item x1: (continuous) legitimate explanatory variable: num  -0.33 -1.22 -1.28 -0.93 0.8 -0.33 0.05 1 -0.07 -1.34 ...
\item x2: (continuous) legitimate explanatory variable: num  7.9 3.7 9.6 4.5 9.2 5.8 2.8 1.2 1.4 3.8 ...
\item x3: (continuous) legitimate explanatory variable: num  -0.23 -1.19 -0.41 -1.05 -0.65 -0.66 -0.89 0.25 -1.53 -0.3 ...
\item score: (continuous) true score used to generate outcomes: 0.46 0.09 0.5 0.13 0.84 0.27 0.14 0.22 0.07 0.09 ...
\item sensitive: (binary) sensistive attribute: chr  "A" "A" "A" "A" ...
\item y: (binary) outome: int  1 0 0 0 1 0 0 0 0 0 ...
}
}
\examples{
data(toydata2)
data(toydata2_train)
data(toydata2_valid)
}
\keyword{dataset}
