% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{frenchmotor}
\alias{frenchmotor}
\title{Motor insurance data}
\format{
data frame
}
\usage{
data(frenchmotor)
}
\description{
Dataset from CASDataset R package
}
\examples{
data(frenchmotor)
str(frenchmotor)
}
\keyword{dataset}
