% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Compute the calibration function}
\usage{
plot_calibration(predy, y, u = seq(0, 1, by = 0.01))
}
\arguments{
\item{predy}{The vector of predicted scores (in $[0,1]$)}

\item{y}{The vector of observed values (in $\{0,1\}$)}

\item{u}{The grid to compute values for the calibration plot (in $[0,1]$)}
}
\value{
A dataframe with $x$ and $y$ used in the calibration plot
}
\description{
Compute values use to plot the calibration plot of a score defined on $[0,1]$
}
\examples{
d1 <- 1
}
